Fix constants
